pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vatavaran"
version = "0.1.0"
description = "Vatavaran OpenEnv environment for cloud root cause analysis."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.2",
  "fastapi>=0.115.0",
  "pydantic>=2.8.0",
  "uvicorn>=0.30.0",
  "requests>=2.32.0",
  "pandas>=2.2.0",
  "ipython>=8.18.0",
  "PyYAML>=6.0.0",
  "tiktoken>=0.7.0",
  "openai>=1.54.0",
  "pytz>=2022.7",
  "streamlit>=1.36.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
server = "vatavaran.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["vatavaran*", "server*"]

[tool.setuptools.package-data]
vatavaran = [
  "config/*.yaml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
